stock-alert-bot/requirements.txt
yfinance
aiohttp
motor
python-dotenv
